% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs.subset.R
\docType{methods}
\name{cpt.locations}
\alias{cpt.locations}
\title{Process changepoint locations}
\arguments{
\item{object}{An S4 object produced by \code{\link{wbs.subset}}.}
}
\value{
A data.frame containing the affected variates and the changepoint locations.
}
\description{
Extracts the changepoint locations and affected variates from the results of \code{\link{wbs.subset}}.
}
