% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.subset_fixed.R
\name{plot.subset_fixed}
\alias{plot.subset_fixed}
\title{SUBspace Shrinkage via Exponential Tilting for estimated
projection matrix}
\usage{
\method{plot}{subset_fixed}(x, include_CI = TRUE, ...)
}
\arguments{
\item{x}{Object of class "subset_gibbs", returned by
the function SUBSET_gibbs().}

\item{include_CI}{logical.}

\item{...}{Additional graphical parameters.}
}
\description{
plot.subset_fixed will plot the transition of each
parameter as v, the shrinkage hyperparameter,
varies.
}
