% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.subset_gibbs.R
\name{plot.subset_gibbs}
\alias{plot.subset_gibbs}
\title{SUBspace Shrinkage via Exponential Tilting for estimated
projection matrix}
\usage{
\method{plot}{subset_gibbs}(
  x,
  type = c("trace", "cummean", "theta")[1],
  burnin,
  include_prior = TRUE,
  plotting_args,
  legend_args,
  prior_args,
  prior_density_args
)
}
\arguments{
\item{x}{Object of class "subset_gibbs", returned by
the function SUBSET_gibbs().}

\item{type}{character denoting traceplots ("trace"),
cumulative mean plots ("cummean"), or histograms
of the model parameters ("theta").}

\item{burnin}{The number of initial samples to disregard before
plotting histograms}

\item{include_prior}{Superimpose the marginal prior
densities on the histograms if type = "theta".}

\item{plotting_args}{List of additional graphical parameters for primary plot.}

\item{legend_args}{List of additional graphical parameters for legend.}

\item{prior_args}{List of additional graphical parameters for superimposed prior plot.}

\item{prior_density_args}{List of arguments to be passed to density() for density estimation.}
}
\description{
plot.subset_gibbs will provide several plotting functions
for an object returned by SUBSET_gibbs() or SUBSET_IS_gibbs().
Obtain traceplots, cumulative mean plots, histograms of
the model parameters, and histogram of the
projection matrix parameter.
}
