% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs.subset.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,capa.class-method}
\title{Displays S4 objects produced by capa methods.}
\usage{
\S4method{show}{wbs.subset.class}(object)
}
\arguments{
\item{object}{An S4 class produced by \code{\link{wbs.subset}}.}
}
\description{
Displays S4 object produced by \code{\link{wbs.subset}}. The information produced is the same as that provided by the summary method.
The method is used by the S4 system for automatic printing.
}
