% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUBSET_fixed.R
\name{SUBSET_fixed}
\alias{SUBSET_fixed}
\title{SUBspace Shrinkage via Exponential Tilting for fixed
projection matrix}
\usage{
SUBSET_fixed(
  Mean0,
  Sigma0,
  prior_draws,
  Proj_mat = Proj(matrix(1, ncol(Sigma0), 1)),
  CI_level = 0.95,
  nu,
  nu_max,
  cl
)
}
\arguments{
\item{Mean0}{vector. The posterior mean corresponding
to the base prior (i.e., untilted prior).}

\item{Sigma0}{matrix. The posterior covariance
matrix corresponding to the base prior. NOTE: The
variable names are extracted from Sigma0, so use
colnames(Sigma0) = ...}

\item{prior_draws}{matrix of prior draws (each row is a draw),
corresponding to the base prior.}

\item{Proj_mat}{matrix.  Projection matrix for the
desired linear subspace you wish to shrink towards.}

\item{CI_level}{numeric between (0,1). Level of the
posterior credible interval.}

\item{nu}{shrinkage parameter.  If missing, the optimal nu via Bayes factors will be used}

\item{nu_max}{maximum value of nu to be considered.}

\item{cl}{optional object of class c("SOCKcluster", "cluster") generated from parallel::makeCluster.}
}
\value{
Object of class "subset_fixed", "subset_asymp_fixed", "data.frame" with the
following structure:
\itemize{
\item variable - Variable names extracted from colnames(Sigma0)
\item mean - Posterior mean
\item sd - Posterior sd
\item \link{(1-CI_level)/2}$\% - lower CI bound
\item \link{1 - (1-CI_level)/2}$\% - upper CI bound
\item v - SUBSET hyperparameter (shrinkage strength)
}
}
\description{
SUBSET_fixed() will compute the asymptotic normal
approximation to the posterior using an exponential
tilted prior that shrinks the  parameters towards a
linear subspace.
}
\details{
SUBSET takes a base prior \eqn{\pi_0} and tilts it towards
a linear subspace via the multiplicative term
\deqn{e^{-\frac{v}{2}\theta'(I-P)\theta}.}
The user must perform the initial Bayesian inference for
this base prior, and compute the posterior mean and covariance.
\emph{It is imperative that the necessary conditions for an asymptotic
normal approximation of the posterior hold for the base prior.}
Then, for a fixed projection matrix \eqn{P}, SUBSET_fixed() will
compute the asymptotic normal distribution corresponding to
the tilted prior.  \code{v_sequence} is a user provided positive
number or vector of positive numbers that control the shrinkage:
higher values imply more shrinkage of the posterior while \eqn{v=0}
corresponds to the base prior with no shrinkage.
}
\examples{
# Simulate where truth lies far from subspace
set.seed(2023)
theta_true = c(1,1.5)
N = 100
ndraws = 1e5
# y ~ N(mu, I_2)
y = mvtnorm::rmvnorm(N,theta_true,diag(2))
# mu ~ N(0,2.5^2 * I_2)
prior_cov = 2.5^2 * diag(2)
# Draw from base prior
prior_draws = 
  mvtnorm::rmvnorm(ndraws, numeric(2), prior_cov)
# Get posterior parameters (under base prior)
Sigma0 = 
  chol2inv(chol( N * diag(2) + chol2inv(chol(prior_cov)) ))
Mean0 = 
  Sigma0 \%*\% diag(2) \%*\% (N * colMeans(y))

tilted_post_off_subspace = 
  SUBSET_fixed(Mean0 = Mean0,
               Sigma0 = Sigma0,
               prior_draws = prior_draws, 
               Proj_mat = Proj(c(1,1)))
tilted_post_off_subspace$nu # The shrinkage parameter selected by Bayes factor
tilted_post_off_subspace$BF_favoring_nu # The Bayes factor favoring nu = tilted_post$nu vs. nu = 0
tilted_post_off_subspace$summary
sum((Mean0 - theta_true)^2) # MSE for posterior mean under base prior
sum((tilted_post_off_subspace$summary$posterior_mean - theta_true)^2) # MSE for posterior mean under subset prior

# Now the truth lies near on the subspace
set.seed(2023)
theta_true = c(1.2,1.25)
N = 100
ndraws = 1e5
# y ~ N(mu, I_2)
y = mvtnorm::rmvnorm(N,theta_true,diag(2))
# mu ~ N(0,2.5^2 * I_2)
prior_cov = 2.5^2 * diag(2)
# Draw from base prior
prior_draws = 
  mvtnorm::rmvnorm(ndraws, numeric(2), prior_cov)
# Get posterior parameters (under base prior)
Sigma0 = 
  chol2inv(chol( N * diag(2) + chol2inv(chol(prior_cov)) ))
Mean0 = 
  Sigma0 \%*\% diag(2) \%*\% (N * colMeans(y))

tilted_post_near_subspace = 
  SUBSET_fixed(Mean0 = Mean0,
               Sigma0 = Sigma0,
               prior_draws = prior_draws, 
               Proj_mat = Proj(c(1,1)))
tilted_post_near_subspace$nu # The shrinkage parameter selected by Bayes factor
tilted_post_near_subspace$BF_favoring_nu # The Bayes factor favoring nu = tilted_post$nu vs. nu = 0
tilted_post_near_subspace$summary
sum((Mean0 - theta_true)^2) # MSE for posterior mean under base prior
sum((tilted_post_near_subspace$summary$posterior_mean - theta_true)^2) # MSE for posterior mean under subset prior

}
