% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUBSET_IS_gibbs.R
\name{SUBSET_IS_gibbs}
\alias{SUBSET_IS_gibbs}
\title{SUBspace Shrinkage via Exponential Tilting for fixed
projection matrix}
\usage{
SUBSET_IS_gibbs(
  draws0,
  pi0_samples,
  P_phi = c("power", "geometric")[1],
  prior_phi,
  phi_sequence = 10,
  v_sequence = seq(0.25, 5, by = 0.25),
  min_ESS = 1/2 * nrow(draws0),
  n_u_values = 100,
  n_draws = nrow(draws0),
  verbose = TRUE,
  cl
)
}
\arguments{
\item{draws0}{matrix. The posterior draws
corresponding to the base prior.}

\item{pi0_samples}{Samples of the model parameters
from the \emph{base prior} (not including phi)}

\item{P_phi}{Either (1) a function taking in a scalar and returning
a projection matrix for the model parameters; or (2) "power"
in which case the projection matrix will correspond to
span((1,1,...)',(1^x,2^x,...)'); or (3) "geometric" in
which case the projection matrix will correspond to
span((1,1,...)',(x^1,x^2,...)').}

\item{prior_phi}{list with named arguments "q" and "d".  prior_phi$q()
should provide the quantile function, and prior_phi$d() should provide
the density function.}

\item{v_sequence}{vector of non-negative values for exponential shrinkage.  Higher
implies more shrinkage.}

\item{min_ESS}{integer.  Minimum number of effective sample size to get
u value. WARNING:
Don't make this too high, or you risk not representing the full parameter space well.}

\item{n_u_values}{Number of values to try for u, the hyperparameter of the
proposal distribution that shrinks towards the linear subspace}

\item{verbose}{logical. Should any output be provided?}

\item{cl}{parallel socket cluster (see parallel::makeCluster()).}
}
\value{
Object of class "subset_gibbs", "subset_is_gibbs", with the following structure:
\itemize{
\item samples - 3-dim array.  First index is the samples,
second index is the variable, and third index is the
value of v (the exponential tilting parameter)
\item acc_rate: numeric giving the acceptance rate for phi
\item draws0
\item pi0_samples
\item P_phi
\item prior_phi
\item phi_pmf - probability mass function for \phi
\item v_sequence
}
}
\description{
SUBSET_IS_gibbs() will obtain posterior inference
using an exponential tilted prior that shrinks the
parameters towards a linear subspace.
}
\details{
SUBSET_IS_gibbs() will compute perform a 2-block Gibbs sampler,
where the blocks correspond to the primary model parameters
and the projection matrix parameter (only one parameter
supported at this time).  The Gibbs algorithm performs importance
sampling to obtain draws from the conditional posterior of theta
using an exponential tilted prior that shrinks the
parameters towards a linear subspace for a given projection
matrix.
\emph{Note that that the parameter space must be convex.} (It usually is.)
}
