% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs.subset.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,wbs.subset.class-method}
\title{Summary of results produced by wbs.subset.}
\usage{
\S4method{summary}{wbs.subset.class}(object)
}
\description{
Summary methods for S4 object returned by \code{wbs.subset}. The output lists the paramaters provided by the user (or the defaults used), the dimensions of the data,
the location of each changepoint, and the affected variates.
}
