% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.subset_gibbs.R
\name{plot.subset_gibbs}
\alias{plot.subset_gibbs}
\title{SUBspace Shrinkage via Exponential Tilting for estimated
projection matrix}
\usage{
\method{plot}{subset_gibbs}(
  x,
  type = c("trace", "cummean", "theta", "phi", "transition")[1],
  v_index,
  include_prior = TRUE,
  ...
)
}
\arguments{
\item{x}{Object of class "subset_gibbs", returned by
the function SUBSET_gibbs().}

\item{type}{character denoting traceplots ("trace"),
cumulative mean plots ("cummean"), histograms
of the model parameters ("theta"), or a histogram
of the projection matrix parameter ("phi"), or
transition plot of the theta values over v ("transition").}

\item{v_index}{Integer giving the "slice" index of
the posterior samples corresponding to the desired
value of v, i.e., the exponential tilting parameter.
Only used for type = "theta" or "phi".}

\item{include_prior}{Superimpose the marginal prior
densities on the histograms if type = "theta" or "phi".}

\item{...}{Additional graphical parameters.}
}
\description{
plot.subset_gibbs will provide several plotting functions
for an object returned by SUBSET_gibbs().  Obtain
traceplots, cumulative mean plots, histograms of
the model parameters, and histogram of the
projection matrix parameter.
}
