% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUBSET_fixed.R
\name{SUBSET_fixed}
\alias{SUBSET_fixed}
\title{SUBspace Shrinkage via Exponential Tilting for fixed
projection matrix}
\usage{
SUBSET_fixed(
  Mean0,
  Sigma0,
  Proj_mat = Proj(matrix(1, ncol(Sigma0), 1)),
  v_sequence = seq(0.25, 5, by = 0.25),
  CI_level = 0.95,
  verbose = FALSE
)
}
\arguments{
\item{Mean0}{vector. The posterior mean corresponding
to the base prior (i.e., untilted prior).}

\item{Sigma0}{matrix. The posterior covariance
matrix corresponding to the base prior. NOTE: The
variable names are extracted from Sigma0, so use
colnames(Sigma0) = ...}

\item{Proj_mat}{matrix.  Projection matrix for the
desired linear subspace you wish to shrink towards.}

\item{v_sequence}{vector of positive values. See details.}

\item{CI_level}{numeric between (0,1). Level of the
posterior credible interval.}

\item{verbose}{logical. Should any output be provided?  Note that
this is typically very fast, so usually not needed.}
}
\description{
SUBSET_fixed() will compute the asymptotic normal
approximation to the posterior using an exponential
tilted prior that shrinks the  parameters towards a
linear subspace.
}
\details{
SUBSET takes a base prior \eqn{\pi_0} and tilts it towards
a linear subspace via the multiplicative term
\deqn{e^{-\frac{v}{2}\theta'(I-P)\theta}.}
The user must perform the initial Bayesian inference for
this base prior, and compute the posterior mean and covariance.
\emph{It is imperative that the necessary conditions for an asymptotic
normal approximation of the posterior hold for the base prior.}
Then, for a fixed projection matrix \eqn{P}, SUBSET_fixed() will
compute the asymptotic normal distribution corresponding to
the tilted prior.  \code{v_sequence} is a user provided positive
number or vector of positive numbers that control the shrinkage:
higher values imply more shrinkage of the posterior while \eqn{v=0}
corresponds to the base prior with no shrinkage.
}
\examples{
post_mean_from_pi0 = c(1,1.5)
post_cov_from_pi0 = diag(c(1,3))
P = Proj(rep(1,2))
SUBSET_fixed(Mean0 = post_mean_from_pi0, Sigma = post_cov_from_pi0, Proj_mat = P, v_sequence = c(0.5,100))

}
