% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUBSET_gibbs_discrete.R
\name{SUBSET_gibbs_discrete}
\alias{SUBSET_gibbs_discrete}
\title{SUBspace Shrinkage via Exponential Tilting for estimated
projection matrix}
\usage{
SUBSET_gibbs_discrete(
  Mean0,
  Sigma0,
  prior_draws,
  n_draws,
  P_phi = c("power", "geometric")[1],
  unique_phi,
  prior_phi,
  nu,
  nu_max,
  verbose = TRUE
)
}
\arguments{
\item{Mean0}{vector. The posterior mean corresponding
to the base prior (i.e., untilted prior).}

\item{Sigma0}{matrix. The posterior covariance
matrix corresponding to the base prior. NOTE: The
variable names are extracted from Sigma0, so use
colnames(Sigma0) = ...}

\item{prior_draws}{Samples of the model parameters
from the \emph{prior} (not including phi)}

\item{n_draws}{positive integer of how many posterior
samples are desired.}

\item{P_phi}{Either (1) a function taking in a scalar and returning
a projection matrix for the model parameters; or (2) "power"
in which case the projection matrix will correspond to
span((1,1,...)',(1^x,2^x,...)'); or (3) "geometric" in
which case the projection matrix will corrspond to
span((1,1,...)',(x^1,x^2,...)').}

\item{unique_phi}{vector of the finite values the phi can take}

\item{prior_phi}{numeric vector of probabilities for unique_phi}

\item{nu}{shrinkage parameter.  If missing, the optimal nu via Bayes factors will be used}

\item{nu_max}{maximum value of nu to be considered.}

\item{verbose}{logical. Should any output be provided?}
}
\value{
Object of class "subset_gibbs", "subset_asymp_gibbs",
with the following structure:
\itemize{
\item samples - matrix of posterior draws from the 2-block Gibbs sampler. Each
row is a posterior draw.  The last column corresponds to phi.
\item acc_rate: numeric giving the acceptance rate for phi
\item Mean0
\item Sigma0
\item P_phi
\item prior_phi
}
}
\description{
SUBSET_gibbs_discrete() is a more computationally efficient
implementation of SUBSET_gibbs() when the values of phi are
finite.  It will compute perform a 2-block Gibbs sampler,
where the blocks correspond to the primary model parameters
and the projection matrix parameter (only one parameter
supported at this time).  The Gibbs algorithm relies on
the asymptotic normal approximation to the conditional posterior
using an exponential tilted prior that shrinks the
parameters towards a linear subspace for a given projection
matrix.
}
\details{
SUBSET takes a base prior \eqn{\pi_0} and tilts it towards
a linear subspace via the multiplicative term
\deqn{e^{-\frac{v}{2}\theta'(I-P)\theta}.}
The user must perform the initial Bayesian inference for
this base prior, and compute the posterior mean and covariance.
\emph{It is imperative that the necessary conditions for an asymptotic
normal approximation of the posterior hold for the base prior.}
For a projection matrix \eqn{P} parameterized by \eqn{\phi},
SUBSET_gibbs() will obtain joint draws of the model parameters and
\eqn{\phi}. \code{v_sequence} is a user provided positive
number or vector of positive numbers that control the shrinkage:
higher values imply more shrinkage of the posterior while \eqn{v=0}
corresponds to the base prior with no shrinkage.
}
