% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUBSET_IS_fixed.R
\name{SUBSET_IS_fixed}
\alias{SUBSET_IS_fixed}
\title{SUBspace Shrinkage via Exponential Tilting for fixed
projection matrix}
\usage{
SUBSET_IS_fixed(
  draws0,
  Proj_mat = Proj(matrix(1, ncol(Sigma0), 1)),
  v_sequence = seq(0.25, 5, by = 0.25),
  CI_level = 0.95,
  min_ESS = 1/2 * nrow(draws0),
  n_u_values = 100,
  verbose = TRUE
)
}
\arguments{
\item{draws0}{matrix. The posterior draws
corresponding to the base prior.}

\item{Proj_mat}{matrix.  Projection matrix for the
desired linear subspace you wish to shrink towards.}

\item{v_sequence}{vector of positive values. See details.}

\item{CI_level}{numeric between (0,1). Level of the
posterior credible interval.}

\item{min_ESS}{integer.  Minimum number of effective sample size. WARNING:
Don't make this too high, or you risk not representing the full parameter space well.}

\item{n_u_values}{Number of values to try for u, the hyperparameter of the
proposal distribution that shrinks towards the linear subspace.}

\item{verbose}{logical. Should any output be provided?}
}
\value{
Object of class "subset_fixed", "subset_IS_fixed", with named elements summary
having the following structure:
\itemize{
\item variable - Variable names extracted from colnames(Sigma0)
\item mean - Posterior mean
\item \link{(1-CI_level)/2}$\% - lower CI bound
\item \link{1 - (1-CI_level)/2}$\% - upper CI bound
\item v - SUBSET hyperparameter (shrinkage strength)
\item ESS - Effective sample size for the importance sampling
\item u - the maximum u providing at ESS >= ESS_min
}
and is_draws, which is a list of, for each v, draws from the importance distribution
and the corresponding importance weights.
}
\description{
SUBSET_IS_fixed() will perform importance sampling
using an exponential tilted prior that shrinks the
parameters towards a linear subspace.
}
\details{
SUBSET takes a base prior \eqn{\pi_0} and tilts it towards
a linear subspace via the multiplicative term
\deqn{e^{-\frac{v}{2}\theta'(I-P)\theta}.}
The user must perform the initial Bayesian inference for
this base prior, and input posterior draws.
Then, for a fixed projection matrix \eqn{P}, SUBSET_IS_fixed() will
perform the importance sampling algorithm corresponding to
the tilted prior.  \code{v_sequence} is a user provided positive
number or vector of positive numbers that control the shrinkage:
higher values imply more shrinkage of the posterior while \eqn{v=0}
corresponds to the base prior with no shrinkage.
\emph{Note that that the parameter space must be convex.} (It usually is.)
}
\examples{
theta_true = c(1,1.5)
N = 50
y = mvtnorm::rmvnorm(N,theta_true,diag(2))
P = Proj(c(1,1))
posterior_Sigma = function(nu){
  chol2inv(chol((N+1) * diag(2) + nu * (diag(2) - P)))
}
posterior_mu = function(nu){
  N * posterior_Sigma(nu) \%*\% colMeans(y)
}
ndraws = 1e4
draws0 = 
  mvtnorm::rmvnorm(ndraws,posterior_mu(0),posterior_Sigma(0))
tilted_post = 
  SUBSET_IS_fixed(draws0,
                  Proj_mat = P,
                  v_sequence = seq(5,50,by = 5),
                  CI_level = 0.95,
                  min_ESS = 1/2 * nrow(draws0),
                  n_u_values = 100,
                  verbose = TRUE)
tilted_post$summary
}
